Relax type checking of path arguments.
Instead of requiring `str` or the specific path type of the context::

  `not isinstance(other, (str, LocalPath))`

guard against the wrong path type for the context::

  `isinstance(other, RemotePath)`

To avoid circular imports between local_machine.py and
remote_machine.py, factor RemotePath into a separate module
remote_path.py.
